ListView now works offline

+++ iOSAvanzado/Presentation/HeroesList/HeroesListViewModel.swift
@@ -4,13 +4,13 @@ import CoreData
 class HeroesListViewModel: HeroesListViewControllerDelegate {
     
     // MARK: - Properties
-    private let database = CoreDataStack()
+    private let database: CoreDataStackProtocol
     
     var loginVieModel: LoginViewControllerDelegate
     var mapViewModel: MapViewControllerDelegate
     var viewState: ((HeroesViewState) -> Void)?
     var heroesCount: Int { heroesList.count }
-
+    
     var heroesList: Heroes = []
     
     
@@ -19,29 +19,35 @@ class HeroesListViewModel: HeroesListViewControllerDelegate {
          mapViewModel: MapViewControllerDelegate) {
         self.loginVieModel = loginViewModel
         self.mapViewModel = mapViewModel
+        self.database = CoreDataStack()
     }
-
+    
     
     // MARK: - Public functions
     func onViewAppear() {
         viewState?(.loading(true))
-
+        
         DispatchQueue.global().async { [weak self] in
             defer { self?.viewState?(.loading(false)) }
-
-            ApiProvider.shared.getHeroes(by: "") { [weak self] heroes in
-                
-                DispatchQueue.main.async { [weak self] in
+            
+            if SecureDataProvider.shared.isLogged {
+                self?.heroesList = self?.database.fetchHeroes() ?? []
+                self?.viewState?(.updateData)
+            } else {
+                ApiProvider.shared.getHeroes(by: "") { [weak self] heroes in
+                    
+                    self?.heroesList = heroes
+                    
                     self?.database.deleteHeroesData()
                     self?.database.saveHeroes(heroes)
+                    self?.viewState?(.updateData)
                 }
                 
-                self?.heroesList = heroes
-                self?.viewState?(.updateData)
             }
         }
     }
-
+    
+    
     func heroBy(index: Int) -> Hero? {
         if index >= 0 && index < heroesCount {
             heroesList[index]
@@ -49,7 +55,7 @@ class HeroesListViewModel: HeroesListViewControllerDelegate {
             nil
         }
     }
-
+    
     func heroDetailViewModel(index: Int) -> HeroDetailViewControllerDelegate? {
         guard let selectedHero = heroBy(index: index) else { return nil }
         
@@ -61,10 +67,10 @@ class HeroesListViewModel: HeroesListViewControllerDelegate {
     
     func onLogoutPressed() {
         // Borramos el token al cerrar sesión
-        DispatchQueue.main.async { [weak self] in
-            self?.database.deleteAllData()
-        }
+        database.deleteAllData()
+        
         
         SecureDataProvider.shared.remove(token: SecureDataProvider.shared.getToken() ?? "")
+        SecureDataProvider.shared.isLogged = false
     }
 }
